rename to just "container-onnxruntime"
