+++ CMakeLists.txt
@@ -88,7 +88,7 @@ add_subdirectory(container-apache-http-client-bundle)
 add_subdirectory(container-core)
 add_subdirectory(container-disc)
 add_subdirectory(container-messagebus)
-add_subdirectory(container-onnxruntime-bundle)
+add_subdirectory(container-onnxruntime)
 add_subdirectory(container-search)
 add_subdirectory(container-search-and-docproc)
 add_subdirectory(container-spifly)
