+++ config-model/src/main/java/com/yahoo/vespa/model/container/ContainerModelEvaluation.java
@@ -28,7 +28,7 @@ public class ContainerModelEvaluation implements
 
     private final static String EVALUATION_BUNDLE_NAME = "model-evaluation";
     private final static String INTEGRATION_BUNDLE_NAME = "model-integration";
-    private final static String ONNXRUNTIME_BUNDLE_NAME = "container-onnxruntime-bundle.jar";
+    private final static String ONNXRUNTIME_BUNDLE_NAME = "container-onnxruntime.jar";
 
     private final static String EVALUATOR_NAME = ModelsEvaluator.class.getName();
     private final static String REST_HANDLER_NAME = "ai.vespa.models.handler.ModelsEvaluationHandler";
