+++ container-dev/pom.xml
@@ -80,7 +80,7 @@
     </dependency>
     <dependency>
       <groupId>com.yahoo.vespa</groupId>
-      <artifactId>container-onnxruntime-bundle</artifactId>
+      <artifactId>container-onnxruntime</artifactId>
       <version>${project.version}</version>
       <exclusions>
         <exclusion>
