+++ container-onnxruntime/.gitignore
@@ -1,0 +1,1 @@
+dependency-reduced-pom.xml
