+++ container-onnxruntime/CMakeLists.txt
@@ -1,2 +1,2 @@
 # Copyright Yahoo. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
-install_jar(container-onnxruntime-bundle.jar)
+install_jar(container-onnxruntime.jar)
