+++ container-onnxruntime/README.md
@@ -1,4 +1,4 @@
 <!-- Copyright Yahoo. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root. -->
-# container-onnxruntime-bundle
+# container-onnxruntime
 
 onnxruntime.ai repackaged as a bundle
