+++ model-integration/pom.xml
@@ -53,7 +53,7 @@
     </dependency>
     <dependency>
       <groupId>com.yahoo.vespa</groupId>
-      <artifactId>container-onnxruntime-bundle</artifactId>
+      <artifactId>container-onnxruntime</artifactId>
       <version>${project.version}</version>
       <scope>provided</scope>
     </dependency>
