+++ pom.xml
@@ -58,7 +58,7 @@
         <module>container-disc</module>
         <module>container-documentapi</module>
         <module>container-messagebus</module>
-        <module>container-onnxruntime-bundle</module>
+        <module>container-onnxruntime</module>
         <module>container-search-and-docproc</module>
         <module>container-search</module>
         <module>container-spifly</module>
